[REF] composer: move 'getTokenColor' from composer to abstract_composer_store
Task: 4123205
